Create DAX-density
The R is used to analyze and visualize the distributional properties of daily stock index returns. Specifically, the implementation involves plotting both the histogram and kernel density estimate to visually present characteristics of the data. The data come from the "EuStockMarkets" dataset in the "datasets" package, focusing on the historical closing prices of the German DAX index. Based on this, daily returns are calculated and used for further analysis and visualization.